requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.9
requests==2.32.3
lxml==5.3.0
python-dotenv==1.0.1
httpx==0.27.0
